Move message logic from user bp to own blueprint
