+++ app/routes/message/__init__.py
@@ -1,0 +1,4 @@
+from flask import Blueprint
+
+bp = Blueprint("message", __name__)
+from app.routes.message import routes
